chop up othe part to calculate the number of rows and columns
